requirements.txt
langchain>=0.1.0
openai>=1.0.0
redis>=5.0.0
python-dotenv>=1.0.0
fastapi>=0.100.0
uvicorn>=0.24.0
pydantic>=2.0.0
SQLAlchemy>=2.0.0
python-multipart>=0.0.6
PyYAML>=6.0.1
pillow>=10.0.0
pytesseract>=0.3.10
whisper>=1.1.10
pandas>=2.0.0
numpy>=1.24.0 